+++ R/ocf.R
@@ -56,11 +56,12 @@
 #' me <- marginal_effects(forests, eval = "atmean")
 #' print(me)
 #' print(me, latex = TRUE)
+#' plot(me)
 #' 
 #' ## Compute standard errors. This requires honest forests.
 #' honest_me <- marginal_effects(honest_forests, eval = "atmean", inference = TRUE)
-#' honest_me$standard.errors
-#' print(honest_me, latex = TRUE)}
+#' print(honest_me, latex = TRUE)
+#' plot(honest_me)}
 #' 
 #' @import utils stats orf
 #' @importFrom Rcpp evalCpp
@@ -70,14 +71,14 @@
 #'
 #' @references
 #' \itemize{
-#'   \item Di Francesco, R. (2023). Ordered Correlation Forest. arXiv preprint \href{https://arxiv.org/abs/2309.08755}{arXiv:2309.08755}.
+#'   \item Di Francesco, R. (2025). Ordered Correlation Forest. Econometric Reviews, 1–17. \href{https://doi.org/10.1080/07474938.2024.2429596}{https://doi.org/10.1080/07474938.2024.2429596}.
 #' }
 #' 
 #' @seealso \code{\link{marginal_effects}}
 #' 
 #' @export
 ocf <- function(Y = NULL, X = NULL,
-                honesty = FALSE, honesty.fraction = 0.5, inference = FALSE, alpha = 0,
+                honesty = FALSE, honesty.fraction = 0.5, inference = FALSE, alpha = 0.2,
                 n.trees = 2000, mtry = ceiling(sqrt(ncol(X))), min.node.size = 5, max.depth = 0, 
                 replace = FALSE, sample.fraction = ifelse(replace, 1, 0.5), n.threads = 1) {
   ## 0.) Defaults for variables not needed.